added BenchmarkingBP, dep parse vars aren't working
for some reason I'm getting NaNs for f->v messages coming
from the dep parse global factor (this is messing a bunch
of things up, so it's been disabled). With dep parse vars
disabled, I needed to add back the overfitting feature to
get the tests to pass, but they do pass.