too many changes... major feature refactors and updates to factors
1. i have basically given up on this bob crap. its unecessary and
breaks easily. i just dropped in a global alphabet, one pass,
but no hopes of doing cool caching stuff.

2. i have disabled the ExpansionHardFactor for f_i and f_i^h
variables. i moved back to making sure everything works before
latent syntax, and i ripped out most of the parts of the model.
i now have easily interchangable features, so i tried to get just
trigger prediction to work with simple features (not real features
with an overfitting feature thrown in). it didn't work, and i
traced the problem back to the expansion hard factors. since we
are choosing frames that have the highest *marginal* likelihood
(i.e. sum out spans in the case of only f_i -- f_i^h factors,
which is what i have now), removing all of the prob mass in the
nullFrame row except for expansion=null means that there is a
strong bias towards choosing a non-null-frame. this is exactly
what i kept seeing: the features would say that you should choose
frame X over nullFrame, but the effect of this feature (basically
a unary factor on f_i, which i promoted to a binary one), was
washed out by the expansion hard factor, which had a lot of
mass missing for nullFrame.

AHH... i think to solve this i need to handle the normalization
of the factors differently. i was considering just eliminating
this hard factor, and proceeding as normal (which might have the
nice side effect of making things faster), but i did originally
have the intuition that we should not include these terms when
normalizing this distribution over (frame,expansion).

whatever the answer, it has implications for how we handle the
frame-role binary factors as well (i.e. the hard factor that
rules out r_ijk=1 for k > f_i.numRoles).