got all the pruning in order, but now i need to fix the arg decoding
tl;dr I want to merge the r_ijk and r_ijk^e variables because
when i'm doing decoding (which up to this commit is wrong), I
really want the marginal distribution for (r_ijk, r_ijk^e),
which I can't get right now. The domain of the new r_ijk variable
would be {null} \cup expansions(j).

For this variable merge to be sucessful, a few things have to
be efficient:
1. Exactly1/AtMost1, which are defined over N binary variables, must
   be generalizable to N k-ary variables, where one of the k values
   corresponds to "off" or "false", as in the binary case.
2. We must be able to afford to sum over phi(f_i, r_ijk, l_ij),
   which is now much bigger.