I fixed the -infinity in the LL by normalizing messages
I sent the following note to myself and Matt:

I discovered that when I normalize messages, I get finite log likelihoods.
The thing that was causing this was the fact that compute LL in two ways based on:

// BeliefPropagation.java
453: public double getPartition() {
454:         if (prm.schedule == BpScheduleType.TREE_LIKE && prm.normalizeMessages == false) {

(for this test I'm still dealing with a tree).

If the condition above holds, you take the product of local Z for every clique.
(I think if messages are not normalized you will still get the correct Z up to a multiplicative constant).

Otherwise, you compute Bethe free energy, which involves expectations of logs of factors,
some of which will have semi-ring 0 in them for me (hard factors).

This may be an issue of 0 * log(0) = 0 vs -Infinity, or it may be something else,
but I just wanted to jot this down so I may revisit it later.