mixed commit: added feature benchmark, but code is generally not working
the pipeline test are not passing having something to do with feature vectors
not being pruned. we are pruning features that don't show up at least K times,
and removing them from the alphabet/model. i want to only compute them once,
so i have to remove them from the vectors themselves, but that doesn't appear
to be working because i'm getting "this feature is not a part of the model"
errors.

independent of this, i was thinking that i should not be predicting/training
the spans of roles at the same time as the role binary vars. basically, it
is twice as many factors/features, but we will necessarily throw away most of
these span features (when we decode a arg/role binary var to be "not realized").
i think i should just put these in a pipeline where we decode the binary vars
first. the other direction is to put a binary factor between the expansion
var and the binary var... but i'm not sure that training that model is feasible.
i'll make a decision one way or another after this commit.