about to refactor ExpansionVar
the problem is that it should be a bunch of binary variables rather than
one N-ary variable. PairedExactly1Factor was meant to handle this, but it
is not needed, may have bugs or otherwise undesireable behaviour (e.g. how
to handle the messages sent to the binary variable -- do you normalize for
how many constituency variables are touching?). it will be much more
understandable and useable if I make it a bunch of binary vars.