found really interesting hack in RoleSpanLabelingStage which boost performance
(only tested on small data, N=200)

the bug/hack means that we throw out a lot of examples which arguably
should be included as negative examples. Not sure if this means that
the model is being overwhelmed by negative examples, or if i've goofed
in the semantics of the variables that i'm rolling out.
- check the gap with larger data
- consider a better way to handle too many negative examples
	- something like ApproxF1MbrDecoder
	- weighted sub-sampling of negative examples