EmbeddingFeatures now supports sparse right features and works great
this was verified by inserting an embedding that is just sparse features
and ensuring that I get the same performance at the end of the day as I
would if I just learned using OldFeaturesParams/Adjoints.Vector

I also added support for saying that your embeddings don't do anything
when your run backwards (for special cases like embeddings that are
features) to save time.

I'm a bit worried about the ability to cache QuadAdjoints when we're
doing a fullUpdate (i.e. using Adjoint caching) and using embeddings.
Back of the envelope:
  Assume: TK=300, #spans=1000, and Adjoints.size=10k*8B
    (10k is for leftTimesRight, which can be skipped if score is given the
    info that the returned Adjoints will never have backwards called on them)
  This is 24 GB per decode, which is not reasonable.
  This sucks because 10k = 200 * 50, which are not huge embeddings.