monster update. feeling the need to checkpoint, even though everything's not done.
Latest realization: I am currently generating prune actions backwards!  I build
the index (to support contains and crosses) on the COMMIT actions when I should
be building it based on the PRUNE actions! I really want to know the set of
COMMIT actions which have a span that is contained/crossed with(in) PRUNE.span,
not the other way around! The nice thing about this mistake is that crossed is
a symmetric relation (contained in is not!), so I didn't accidentally screw
myself over big time with this forced flip.