I now have things relatively stable again.
I think there are a couple issues that lead to things falling apart:
1) global features may be broken (I don't think they were ever carefully tested after the efficiency updates)
2) pretrain (getStatelessUpdate) is certainly broken.

Right now it works if you force it to do full training (see that
Reranker.hasStatefulFeatures is forced to return true!).

testOnTrain=true
1125997  INFO            RerankerTrainer - [tuneModelForF1] chose recallBias=2.0 with ArgumentMicroF1=0.7768860353130016
[main]: ArgumentMicroF1 = 0.796
[main]: ArgumentMicroPRECISION = 0.700
[main]: ArgumentMicroRECALL = 0.921

testOnTrain=false
1528348  INFO            RerankerTrainer - [tuneModelForF1] chose recallBias=0.5 with ArgumentMicroF1=0.17763157894736842
[main]: ArgumentMicroF1 = 0.237
[main]: ArgumentMicroPRECISION = 0.390
[main]: ArgumentMicroRECALL = 0.170

Both of these are with n=100 and no global features. I believe they could both
come up with a bit of tuning, but they're in the right ballpark.