i've realized a fairly big mistake.
The way I have been training is to satisfy a vast subset of the constraints
that I said I would satisfy. Specifically I only enforce margin constraints on
z,y that are *final*, that is z corresponds to a complete decode that specifies
a single parse. I need to satisfy constraints on all the other z as well!

I arrived at this while debugging by giving the model CheatingParams and
observing the decoding decisions it makes. The model learns weights that have
the correct sign (even though I struggled with this for pretrain -- this is
only talking about full train), but their magnitudes are way off. After
training, we would get:

Weights:
                 intercept   -1.735
       currentActionIsGold   +1.253
  pairOfActionsAreBothGold   +2.547

And on the first step of decoding, the first choice it would make:
  s(COMMIT(t,k,s))                = -0.48   # for s==goldArg for (t,k), other s score lower
  s(PRUNE(t,k))                   = +6.17
  deltaLoss(PRUNE(t,k)) = costFN  = +3.00

Which would result in a PRUNE-everything decode. But it is clear that already,
getting this decision right shold correspond to a margin constraint (this is a
z after all!).

Now, the issue is that I realize that my use of beam search is kind of whack. I
pop states off of the beam and then enqueue their children. This gaurantees
that I make progress through the space, but also gaurantees that a z that is
non-terminal will not be the last thing on the beam. This is the way that I was
implicitly only taking the intersection of the possible z's for margin
constraints with those that are terminal.

NOTE: I think I also had the problem that both the oracle and mostViolated
searches could open a node twice, since there was no book-keeping concerning
which states had been explored. This would only show itself if the beam size
was greater than 1 though, because otherwise there is no way to get loops,
you're just doing an argmax out of the current best state.

So, now I need to search for *any z*, not just the terminal ones. In this case,
the loop problem seems even worse. Keeping track of the "states explored" boils
down to keeping sets of BitSets (remember that z are in bijection with the
BitSet, which represents a set of y). I don't think there is an efficient way
to keep track of this, even with the use of stuff like Bloom filters (which
admit false positives, which is not acceptable in this case -- we need to know
for sure if we have seen a state or not). My best solution for this so far is
very clumsy: use the search proceedure that you had before, ignoring the states
explored, and every time you pop an (incomplete) z off of the beam, try to push
it onto an auxilliary beam which tracks the max over the entire explored area.

For this to be acceptable, I need to re-create the proof that I had earlier:
  (z,y) \in R(S_{\theta}, B, x) => (z,y) \in R(S_{\theta} + deltaLoss, B, x)
I believe this still holds.

What does it mean to enforce a constraint on a non-terminal z, which does not
correspond to a single parse/loss?  Non-terminal z must follow MANY
constraints! One for each {y} that is still reachable!  We can prove that if we
satisfy the constraint for the y with the maximal loss, then we've satisfied it
for all y (since they share a z). This is shaping up to look like we either
can't write features on y, or we must write features on sets of y... which is
the same thing as featurizing f(x,z). This is really all you could hope for,
these are the global features on the prefixes/pieces of a decode built up so
far. See my paper for notes on the benefits of this "single order over decode"
style compared to "all at once decode" strategies typical of graph/chart-based
methods.