First big whack at multi-threaded load and serve architecture for CachedFeatures
Previously you had to wait for ~3 hours for CachedFeatureParams to load every
feature from every file (taking up to 60 GB). There was no coordination between
this code and the code that served up FNParses, so there was no way to know for
which FNParses the features had been loaded. This made testing without waiting
the full ~3 hours (and ~60 GB memory) impossible. Additionally, learning would
be very slow to start.

Now, there is coordination between the ItemProvider, the Params (which has now
been refactored out into an inner class), and the Inserter (which runs on its
own thread).

TODO
- Put this ItemProvider in RerankerTrainer instead of the naive one currently
- add locking around loadedItems and have multiple threads of Inserters!