basic Uberts example works.
CAUTION: The way a relation's HypEdge's head node is constructed greatly
affects the semantics of how the transition system works.  For example: I had
set the head node to be the same node for all HypEdges who have the same
Relation (POS in this case). The problem with this was that this meant that all
POS tags would match the graph fragment:
  ROOT -> pos:NodeType -> tokenIndex:NodeType
Since the search was initiated from witness-pos:HypNode, which was adjacent to
every pos tagging. This meant that even though the AtMost1 factor was working
properly and removing the correct set of (not-chosen) POS tags, they would all
be added back because of the over-firing `pos(i,*) => pos(i+1,*)` rule just
described.