improved AtMost1, ner mutual exclusion works, fixed bug in UbertsTests, spans are now [inclusive,exclusive)
If [inclusive,inclusive] spans are used, it is hard to specify graph fragments
for e.g.
  ner(i,j,T) ^ diff(T,S) => !ner(i,j,S)
When you try to match the LHS graph fragment, you use a new TKey(tokenIndex)
for both i and j, both expect to match a new node for the first time. If you
have an ner(5,5,PER), this pattern won't match because 5 and 5 are the same
tokenIndex HypNode. If you use [inclusive,exclusive), you get an easy fix.

There will be cases where you want to use this [inclusive,exclusive) scheme
with tokenIndex nodes to your advantage, e.g.
  constituent(i,j,NP) ^ constituent(j,k,PP) => constituent(i,k,NP)
Which will allow you to write the (sub-)spanning tree of the LHS as
  c1:constituent -> i:tokenIndex -> GOTO_PARENT -> j:tokenIndex -> c2:constituent -> k:tokenIndex
And then use c1,c2 to look up matching grammar rules (NP->NP PP in this case)

If I hadn't used that trick, the other way to fix the 5,5 problem is to have
one NodeType for start and another for end. This is arguably even cleaner,
but prohibits the type of trick described above.