TypeInference now works on values too!
It is basically going to replace the backwards parser.
The major limitation of TransitionGeneratorBackwardsParser that
TypeInference doesn't have is that the former will break if all
the LHS values are not immediately present in the RHS fact, e.g.
  fact: srl4(3,4,propbank/kill-v-01,4,7,ARG1)
  rule: event1'(e1,ts,te) & event2'(e2,e1,f) & srl1'(s1,ss,se) & srl2'(s2,s1,e1) & srl3'(s3,s2,e2,k) => srl4(ts,te,f,ss,se,k)

TransitionGeneratorBackwardsParser doesn't know how to make event2(e1,f)
since e1 can only be constructed by making event1(ts,te) first, and then
using the head node. TypeInference can do this.