Fixed bug in DecisionFunction.ByGroup, e.g. AT_MOST_ONE
Found a case where previous bad assumption that edges which have
DecisionFunction.decide(HypEdge,Adjoints) called on them are always added to
the state. This was based on the idea that edge scores are in
non-ascending-score-order, which is only trivially true (for snapshots of the
agenda, but comparing between time slots is problematic).  For example, adding
an srl3 edge may add many argument4 edges, which will appear on the agenda in
order of score.  Its possible that all of those argument4 edges will be pruned
but they will have decide called on them and appear as if they were added to
the state (or at least have a score as high as any subsequent argument4 edge
that you might add).  Its possible that after those argument4 edges are pruned,
another srl3 edge could be added, thus creating new argument4 edges which have
no strict score relationship to the previous edges (unless we use BFS loop
order).

The new implementation uses a listener to tell when an edge has indeed been
added to the State.

This reveals a slight bug in the semantics of these constraints, they are not
really AT_LEAST_ONE/EXACTLY_ONE/AT_MOST_ONE, they are really FIRST_ONE, which
depends heavily on loop order.