Changed MAX_VIOLATION to only update on Commit actions
Previously I was using a sign flip for Prune actions, encouraging their score to be
as negative as possible. This worked for local models (same perf as DAGGER1) but not global ones.
The reason, as far as I can tell, is that most updates were on Prune actions, and those features
did not cancel out between ORACLE and LOSS_AUGMENTED (they do however cancel out with local models).
I don't have a good counter example for why you shouldn't update on negative examples other than
to rely on my earlier first-principles claim that the score of Prunes is 0.
I think put in more colloquial terms, once a given fact has won out in a group, there is no need
to lower all of the other facts in the group; the real signal for learning is in the Committed
facts chosen by ORACLE and LOSS_AUGMENTED.

I'm going to try this out now in a full-scale experiment.